# apps/drive/public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Xenode Internal API",
    "description": "🔒 **INTERNAL USE ONLY** — This API reference is for internal development and integration purposes. All endpoints require authentication via session cookie or API key bearer token.\n\nXenode is a zero-knowledge encrypted cloud storage platform. The backend never sees raw file contents or unencrypted keys. All E2EE operations (encrypt, decrypt, key wrapping) are performed client-side.",
    "version": "0.1.0",
    "contact": {
      "name": "Xenode Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
...
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "paths": {
...
    "/api/objects/{id}/content": {
      "get": {
        "tags": [
          "Objects"
        ],
        "summary": "Issue direct ciphertext URLs for current content or a retained version",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Direct ciphertext URL manifest"
          },
          "404": {
            "description": "Not found"
          }
        },
        "description": "Authorizes the Drive Space and returns objectId, optional url or ordered chunkUrls. No file bytes pass through Next. URLs live at most 300 seconds and no longer than the ProductSession. Response is private, no-store. Multipart retained versions use the same manifest. Range requests go directly to R2."
      }
    },
...
    "/api/share/{token}/download": {
      "post": {
        "tags": [
          "Link Shares"
        ],
        "summary": "Download shared file",
        "description": "Validates access (password if required) and returns a signed download URL. For a bundle, omitted itemId selects the first surviving item. Returns direct R2 GET URLs lasting at most 300 seconds, rounded down to the link expiry. Responses are private, no-store. Downloads atomically consume the allowance; previews do not consume it.",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download URL returned"
          },
          "404": {
            "description": "Share not found"
          }
        }
      }
    },
    "/api/share/{token}/stream": {
      "post": {
        "tags": [
          "Link Shares"
        ],
        "summary": "Stream shared file",
        "description": "Returns a streaming URL for preview/playback of shared content. Returns direct R2 GET URLs lasting at most 300 seconds, rounded down to the link expiry. Responses are private, no-store. Downloads atomically consume the allowance; previews do not consume it.",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream URL returned"
          }
        }
      }
    },
...
    "/api/direct-shares/{id}/download": {
      "get": {
        "tags": ["Direct Shares"],
        "summary": "Download direct-shared file",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Download URL" }
        }
      }
    },
    "/api/direct-shares/{id}/stream": {
      "get": {
        "tags": ["Direct Shares"],
        "summary": "Stream direct-shared file",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Stream URL" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminCookieAuth": {
        "type": "apiKey", "in": "cookie", "name": "__Host-Xenode_admin_session",
        "description": "Drive Admin JWT cookie, validated against the current Admin record and session version. Development uses Xenode_admin_session. This is separate from user authentication."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Generate keys via `POST /api/keys`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie managed by better-auth. 7-day expiry, auto-refreshes every 24h."
      },
      "cronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cron job secret token (CRON_SECRET env var)."
      }
    },
...
  }
}